# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users": {
      "post": {
        "operationId": "UserController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserEntity"
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ]
      },
      "get": {
        "operationId": "UserController_findAllUsers",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "continuationToken",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/api/users/profile": {
      "get": {
        "operationId": "UserController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserEntity"
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ]
      },
      "patch": {
        "operationId": "UserController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserEntity"
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/api/users/username/{name}": {
      "get": {
        "operationId": "UserController_getUserByName",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserEntity"
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/api/users/search": {
      "get": {
        "operationId": "UserController_searchUsers",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "communityId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserEntity"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "UserController_getUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserEntity"
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/api/users/admin/list": {
      "get": {
        "operationId": "UserController_findAllUsersAdmin",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number",
              "default": 50,
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "continuationToken",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "banned",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "OWNER",
                "USER"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/api/users/admin/{id}": {
      "get": {
        "operationId": "UserController_getUserByIdAdmin",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserEntity"
                }
              }
            }
          }
        },
        "summary": "Get a single user with admin-level details",
        "tags": [
          "User"
        ]
      },
      "delete": {
        "operationId": "UserController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          }
        },
        "summary": "Delete a user account (admin action)",
        "tags": [
          "User"
        ]
      }
    },
    "/api/users/admin/{id}/role": {
      "patch": {
        "operationId": "UserController_updateUserRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserEntity"
                }
              }
            }
          }
        },
        "summary": "Update a user's instance role (OWNER/USER)",
        "tags": [
          "User"
        ]
      }
    },
    "/api/users/admin/{id}/ban": {
      "patch": {
        "operationId": "UserController_setBanStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetUserBanStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserEntity"
                }
              }
            }
          }
        },
        "summary": "Ban or unban a user",
        "tags": [
          "User"
        ]
      }
    },
    "/api/users/block/{userId}": {
      "post": {
        "operationId": "UserController_blockUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ]
      },
      "delete": {
        "operationId": "UserController_unblockUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          }
        },
        "summary": "Unblock a user",
        "tags": [
          "User"
        ]
      }
    },
...
    "/api/users/blocked/{userId}": {
      "get": {
        "operationId": "UserController_isUserBlocked",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockedStatusResponseDto"
                }
              }
            }
          }
        },
        "summary": "Check if a specific user is blocked",
        "tags": [
          "User"
        ]
      }
    },
...
    "/api/roles/instance/{roleId}/users/{userId}": {
      "delete": {
        "operationId": "RolesController_removeInstanceRole",
        "parameters": [
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Remove an instance role from a user",
        "tags": [
          "Roles"
        ]
      }
    },
...
    "/api/membership/user/{userId}": {
      "get": {
        "operationId": "MembershipController_findAllForUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MembershipResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Membership"
        ]
      }
    },
...
    "/api/membership/community/{communityId}/user/{userId}": {
      "get": {
        "operationId": "MembershipController_findOne",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "communityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Membership"
        ]
      },
      "delete": {
        "operationId": "MembershipController_remove",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "communityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Membership"
        ]
      }
    },
...
    "/api/channel-membership/user/{userId}": {
      "get": {
        "operationId": "ChannelMembershipController_findAllForUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChannelMembershipResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "ChannelMembership"
        ]
      }
    },
...
    "/api/channel-membership/channel/{channelId}/user/{userId}": {
      "get": {
        "operationId": "ChannelMembershipController_findOne",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChannelMembershipResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "ChannelMembership"
        ]
      },
      "delete": {
        "operationId": "ChannelMembershipController_remove",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "channelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "ChannelMembership"
        ]
      }
    },
...
    "/api/presence/user/{userId}": {
      "get": {
        "operationId": "PresenceController_getUserPresence",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPresenceResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Presence"
        ]
      }
    },
...
    "/api/presence/users/{userIds}": {
      "get": {
        "operationId": "PresenceController_getMultipleUserPresence",
        "parameters": [
          {
            "name": "userIds",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkPresenceResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Presence"
        ]
      }
    },
...
    "/api/storage/users/{userId}": {
      "get": {
        "operationId": "StorageQuotaController_getUserStorageStats",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserStorageStatsDto"
                }
              }
            }
          }
        },
        "summary": "Get specific user's storage stats (admin only)",
        "tags": [
          "StorageQuota"
        ]
      }
    },
...
  },
  "info": {
    "title": "Kraken API",
    "description": "Kraken Chat Application API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "username": {
            "type": "string",
            "minLength": 2,
            "maxLength": 32
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128
          },
          "email": {
            "type": "string",
            "format": "email"
          }
        },
        "required": [
          "code",
          "username",
          "password"
        ]
      },
      "UserEntity": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "USER"
            ]
          },
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "bannerUrl": {
            "type": "string",
            "nullable": true
          },
          "lastSeen": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "bio": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "role",
          "id",
          "username",
          "avatarUrl",
          "bannerUrl",
          "lastSeen",
          "displayName",
          "bio",
          "status"
        ]
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 32
          },
          "avatar": {
            "type": "string"
          },
          "banner": {
            "type": "string"
          },
          "bio": {
            "type": "string",
            "maxLength": 500
          },
          "status": {
            "type": "string",
            "maxLength": 128
          }
        }
      },
      "UserListResponseDto": {
        "type": "object",
        "properties": {
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserEntity"
            }
          },
          "continuationToken": {
            "type": "string"
          }
        },
        "required": [
          "users"
        ]
      },
      "AdminUserEntity": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "USER"
            ]
          },
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "verified": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "bannerUrl": {
            "type": "string",
            "nullable": true
          },
          "lastSeen": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "bio": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "statusUpdatedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "banned": {
            "type": "boolean"
          },
          "bannedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "bannedById": {
            "type": "string",
            "nullable": true
          },
          "storageQuotaBytes": {
            "format": "int64",
            "type": "integer"
          },
          "storageUsedBytes": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "role",
          "id",
          "username",
          "email",
          "verified",
          "createdAt",
          "avatarUrl",
          "bannerUrl",
          "lastSeen",
          "displayName",
          "bio",
          "status",
          "statusUpdatedAt",
          "banned",
          "bannedAt",
          "bannedById",
          "storageQuotaBytes",
          "storageUsedBytes"
        ]
      },
      "AdminUserListResponseDto": {
        "type": "object",
        "properties": {
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminUserEntity"
            }
          },
          "continuationToken": {
            "type": "string"
          }
        },
        "required": [
          "users"
        ]
      },
      "UpdateUserRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "USER"
            ]
          }
        },
        "required": [
          "role"
        ]
      },
      "SuccessResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ]
      },
      "BlockedStatusResponseDto": {
        "type": "object",
        "properties": {
          "blocked": {
            "type": "boolean"
          }
        },
        "required": [
          "blocked"
        ]
      },
...
      "UserPresenceResponseDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "isOnline": {
            "type": "boolean"
          }
        },
        "required": [
          "userId",
          "isOnline"
        ]
      },
      "BulkPresenceResponseDto": {
        "type": "object",
        "properties": {
          "presence": {
            "type": "object",
            "additionalProperties": {
              "type": "boolean"
            }
          }
        },
        "required": [
          "presence"
        ]
      },
...
      "SetUserBanStatusDto": {
        "type": "object",
        "properties": {
          "banned": {
            "type": "boolean"
          }
        },
        "required": [
          "banned"
        ]
      },
...
    }
  }
}